extract preparsing to uploadtask
